feat(repo/dashboard ): include list of detected package files
 - divide dependencies by base branches
